% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{calc_multinom_deviance}
\alias{calc_multinom_deviance}
\title{Calculate Multinomial Deviance for a Fitted Case-Base Model}
\usage{
calc_multinom_deviance(cb_data, fit_object, all_event_levels)
}
\arguments{
\item{cb_data}{A case-base dataset from \code{create_cb_data}.}

\item{fit_object}{The output from \code{fit_cb_model}.}

\item{all_event_levels}{A vector of all possible event levels to ensure
dimensional consistency.}
}
\value{
The multinomial deviance value.
}
\description{
Calculate Multinomial Deviance for a Fitted Case-Base Model
}
\keyword{internal}
