% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.r
\name{calculate_multinomial_loss}
\alias{calculate_multinomial_loss}
\title{Calculate Multinomial Negative Log-Likelihood (Loss)}
\usage{
calculate_multinomial_loss(coefficients, X, Y, offset = NULL)
}
\arguments{
\item{coefficients}{A p x K matrix of model coefficients.}

\item{X}{A n x p design matrix of predictors.}

\item{Y}{A numeric vector of length 'n' containing the true class labels.
The baseline class should be coded as 0.}

\item{offset}{A numeric vector of length 'n' for the offset term.}
}
\value{
The total negative log-likelihood (a single numeric value).
}
\description{
This function computes the total negative log-likelihood for a multinomial
logistic regression model with a baseline class.
}
