% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.r
\name{calculate_penalized_multinomial_loss}
\alias{calculate_penalized_multinomial_loss}
\title{Calculate Penalized Multinomial Negative Log-Likelihood}
\usage{
calculate_penalized_multinomial_loss(
  coefficients,
  X,
  Y,
  offset = NULL,
  lambda,
  alpha,
  penalty_weights = NULL
)
}
\arguments{
\item{coefficients}{A p x K matrix of model coefficients. Assumes the first
row corresponds to the non-penalized intercept.}

\item{X}{A n x p design matrix.}

\item{Y}{A numeric vector of length n with class labels (0 for baseline).}

\item{offset}{A numeric vector of length n for the offset term.}

\item{lambda}{A single numeric value for the overall regularization strength.}

\item{alpha}{The Elastic Net mixing parameter (1 for Lasso, 0 for Ridge).}

\item{penalty_weights}{A numeric vector of weights for each penalized covariate,
of length p-1. Defaults to 1 for all variables.}
}
\value{
The total penalized negative log-likelihood.
}
\description{
Adds the Elastic Net penalty to the negative log-likelihood loss.
}
