% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{create_lambda_grid}
\alias{create_lambda_grid}
\title{Generate a Lambda Grid for Regularization}
\usage{
create_lambda_grid(
  cb_data,
  lambda,
  nlambda,
  lambda_max,
  lambda.min.ratio,
  regularization,
  alpha,
  n_unpenalized,
  ...
)
}
\arguments{
\item{cb_data}{Case-base data.}

\item{lambda}{A user-supplied lambda sequence.}

\item{nlambda}{Number of lambda values.}

\item{lambda_max}{The maximum lambda value.}

\item{lambda.min.ratio}{Ratio for smallest lambda.}

\item{regularization}{Type of regularization.}

\item{alpha}{The alpha value.}

\item{n_unpenalized}{Number of unpenalized predictors.}

\item{...}{Additional arguments.}
}
\value{
A numeric vector of lambda values.
}
\description{
Generate a Lambda Grid for Regularization
}
