% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv_cb_model}
\alias{cv_cb_model}
\title{Cross-Validation for Penalized Multinomial Case-Base Models}
\usage{
cv_cb_model(
  formula,
  data,
  regularization = "elastic-net",
  alpha = 0.5,
  nfold = 5,
  nlambda = 50,
  ncores = parallel::detectCores()/2,
  fit_fun = mtool::mtool.MNlogistic,
  ratio = 20,
  ratio_event = 1,
  ...
)
}
\value{
An object of class \code{cb.cv} containing the results.
}
\description{
Cross-Validation for Penalized Multinomial Case-Base Models
}
