% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv_cbSCRIP}
\alias{cv_cbSCRIP}
\title{Cross-Validation for Penalized Multinomial Case-Base Models}
\usage{
cv_cbSCRIP(
  formula,
  data,
  regularization = "elastic-net",
  cb_data = NULL,
  alpha = NULL,
  lambda = NULL,
  nfold = 5,
  nlambda = 50,
  ncores = parallel::detectCores()/2,
  n_unpenalized = 2,
  ratio = 50,
  ratio_event = "all",
  lambda_max = NULL,
  lambda.min.ratio = NULL,
  warm_start = TRUE,
  select = c("1se", "min"),
  optimizer = c("CCD", "SAGA", "SVRG", "FISTA"),
  ...
)
}
\arguments{
\item{formula}{A formula object.}

\item{data}{The input data.frame.}

\item{regularization}{Type of regularization.}

\item{cb_data}{A case-base data list.}

\item{alpha}{The alpha value.}

\item{lambda}{A vector of lambda values.}

\item{nfold}{Number of CV folds.}

\item{nlambda}{Number of lambda values.}

\item{ncores}{Number of cores for parallel execution.}

\item{n_unpenalized}{Number of unpenalized predictors.}

\item{ratio}{Sampling ratio.}

\item{ratio_event}{Event ratio.}

\item{lambda_max}{Max lambda.}

\item{lambda.min.ratio}{Min lambda ratio.}

\item{warm_start}{Logical, whether to use warm starts.}

\item{select}{Character string specifying whether the final model is fit using \code{"1se"} (1-standard-error rule) or \code{"min"} (minimum deviance). Defaults to \code{"1se"}.}

\item{optimizer}{The solver to use: one of "CCD", "FISTA", "SAGA" or "SVRG".}

\item{...}{Additional arguments.}
}
\value{
An object of class \code{cb.cv} containing the results.
}
\description{
Cross-Validation for Penalized Multinomial Case-Base Models
}
