% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{find_lambda_max}
\alias{find_lambda_max}
\title{Find the Maximum Lambda (lambda_max)}
\usage{
find_lambda_max(cb_data, n_unpenalized, alpha)
}
\arguments{
\item{cb_data}{Case-base data list.}

\item{n_unpenalized}{Number of unpenalized predictors.}

\item{alpha}{The alpha value for elastic net.}
}
\value{
The estimated lambda_max value.
}
\description{
Find the Maximum Lambda (lambda_max)
}
\keyword{internal}
