% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.r
\name{fit_cb_model}
\alias{fit_cb_model}
\title{Fit a Penalized Multinomial Model on Case-Base Data}
\usage{
fit_cb_model(
  cb_data,
  regularization = c("elastic-net", "SCAD"),
  lambda,
  alpha = NULL,
  optimizer = c("CCD", "FISTA", "SAGA", "SVRG"),
  param_start = NULL,
  n_unpenalized = 2,
  standardize = TRUE,
  all_event_levels = NULL,
  lam1_prev = 0,
  ...
)
}
\arguments{
\item{cb_data}{A case-base dataset from \code{create_cb_data}.}

\item{regularization}{A string, either 'elastic-net' or 'SCAD'.}

\item{lambda}{The primary shrinkage parameter.}

\item{alpha}{The mixing/shape parameter. See \code{prepare_penalty_params}.}

\item{param_start}{Optional starting values for the coefficients.}

\item{n_unpenalized}{Integer. The number of leading covariates to leave unpenalized.}

\item{standardize}{Logical. If TRUE, covariates are scaled to have mean 0 and SD 1.}

\item{all_event_levels}{Optional vector of all event levels.}

\item{...}{Additional arguments passed to the fitting function.}

\item{fit_fun}{The fitting function from \code{cbSCRIP} to use.}
}
\value{
The fitted model object from the specified \code{fit_fun}.
}
\description{
A wrapper function to fit a penalized model (Elastic-Net or SCAD) using
an optimizer from the \code{cbSCRIP} package.
}
