% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.r
\name{fit_cb_path_internal}
\alias{fit_cb_path_internal}
\title{Internal Function to Fit a Regularization Path}
\usage{
fit_cb_path_internal(
  cb_data,
  lambdas,
  regularization,
  alpha,
  n_unpenalized,
  fit_fun,
  param_start = NULL,
  standardize = TRUE,
  all_event_levels = NULL,
  ...
)
}
\arguments{
\item{cb_data}{Case-base data list.}

\item{lambdas}{Vector of lambda values.}

\item{regularization}{Regularization type.}

\item{alpha}{Alpha value.}

\item{n_unpenalized}{Number of unpenalized covariates.}

\item{fit_fun}{Fitting function.}

\item{param_start}{Initial parameters.}

\item{standardize}{Whether to standardize.}

\item{all_event_levels}{Event levels.}

\item{...}{Additional arguments.}
}
\value{
A list of model fits.
}
\description{
Standardizes data once and fits the model for a sequence of lambdas using warm starts.
}
\keyword{internal}
