% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{MNlogisticCCD}
\alias{MNlogisticCCD}
\title{Multinomial Logistic Regression with CCD}
\usage{
MNlogisticCCD(
  X,
  Y,
  offset,
  N_covariates,
  regularization = "l1",
  lambda1,
  lambda2 = 0,
  lambda3 = 0,
  pos = FALSE,
  tolerance = 1e-06,
  maxit = 1000,
  ncores = -1,
  group_id = NULL,
  group_weights = NULL,
  groups = NULL,
  groups_var = NULL,
  own_variables = NULL,
  N_own_variables = NULL,
  param_start = NULL,
  verbose = FALSE,
  save_history = FALSE,
  lam1_prev = 0
)
}
\arguments{
\item{X}{Input matrix.}

\item{Y}{Response vector.}

\item{offset}{Offset vector.}

\item{N_covariates}{Number of covariates.}

\item{regularization}{Regularization type.}

\item{lambda1}{Primary penalty parameter.}

\item{lambda2}{Secondary penalty parameter.}

\item{lambda3}{Tertiary penalty parameter.}

\item{pos}{Positivity constraint.}

\item{tolerance}{Convergence tolerance.}

\item{maxit}{Maximum iterations.}

\item{ncores}{Number of cores.}

\item{group_id}{Group IDs.}

\item{group_weights}{Group weights.}

\item{groups}{Groups matrix.}

\item{groups_var}{Groups variable matrix.}

\item{own_variables}{Own variables.}

\item{N_own_variables}{Number of own variables.}

\item{param_start}{Starting parameters.}

\item{verbose}{Verbosity.}

\item{save_history}{Save history.}
}
\value{
A list with model results.
}
\description{
Fits a multinomial logistic regression model using Coordinate Descent (CCD).
}
