% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{path_cbSCRIP}
\alias{path_cbSCRIP}
\title{Fit a Penalized Multinomial Model over a Regularization Path}
\usage{
path_cbSCRIP(
  formula,
  data,
  regularization = "elastic-net",
  cb_data = NULL,
  alpha = NULL,
  lambda = NULL,
  nlambda = 100,
  n_unpenalized = 2,
  ratio = 25,
  ratio_event = 1,
  lambda_max = NULL,
  lambda.min.ratio = NULL,
  warm_start = T,
  ...
)
}
\value{
An object of class \code{cb.path} containing coefficient paths.
}
\description{
Fits the case-base model for a sequence of lambda values using warm starts,
returning the path of coefficients. Includes a progress bar.
}
