% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plot.cbSCRIP.cv}
\alias{plot.cbSCRIP.cv}
\title{Plot Cross-Validation Results}
\usage{
\method{plot}{cbSCRIP.cv}(x, ...)
}
\arguments{
\item{x}{An object of class \code{cbSCRIP.cv}.}

\item{...}{Additional arguments (not currently used).}
}
\value{
A ggplot object.
}
\description{
Plot Cross-Validation Results
}
