% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plot.cbSCRIP.path}
\alias{plot.cbSCRIP.path}
\title{Plot Coefficient Paths from a cb.path Object}
\usage{
\method{plot}{cbSCRIP.path}(x, plot_intercept = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{cb.path}.}

\item{plot_intercept}{Logical. Whether to include the intercept in the plot.}

\item{...}{Not used.}
}
\value{
A ggplot object.
}
\description{
S3 method to plot the regularization path of coefficients.
}
