% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict_CompRisk}
\alias{predict_CompRisk}
\title{Predict Log-Hazard Ratios for a CompRisk Object}
\usage{
predict_CompRisk(object, newdata = NULL)
}
\arguments{
\item{object}{A fitted model object.}

\item{newdata}{A \code{data.frame} in which to look for variables with which to predict.}
}
\value{
A matrix of linear predictors.
}
\description{
This function calculates the linear predictors (log-hazard ratios relative
to the baseline) for a competing risks model.
}
