% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predictEventProb.twoCSlassos}
\alias{predictEventProb.twoCSlassos}
\title{Predict Cause-Specific Event Probabilities (Cumulative Incidence)}
\usage{
predictEventProb.twoCSlassos(
  object,
  newdata,
  times,
  cause,
  lambdavecs,
  indices,
  ...
)
}
\arguments{
\item{object}{A fitted object of class \code{twoCSlassos}.}

\item{newdata}{A data frame for which to predict.}

\item{times}{A numeric vector of times to predict at.}

\item{cause}{The cause of interest.}

\item{lambdavecs}{A list of lambda vectors (from the object).}

\item{indices}{A numeric vector (length 2) of the specific lambda
indices to use for prediction.}

\item{...}{Not used.}
}
\value{
A matrix of cumulative incidence probabilities.
}
\description{
Calculates the cumulative incidence for a specific cause from a
\code{twoCSlassos} object, based on the cause-specific hazards.
}
