% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predictRisk.CompRisk}
\alias{predictRisk.CompRisk}
\title{Predict Absolute Risk for a CompRisk Object}
\usage{
\method{predictRisk}{CompRisk}(object, newdata, times, cause, ...)
}
\arguments{
\item{object}{A model object of class \code{CompRisk}.}

\item{newdata}{A \code{data.frame} containing the predictor variables.}

\item{times}{A numeric vector of time points at which to predict risk.}

\item{cause}{The specific event type for which to predict the absolute risk.}

\item{...}{Additional arguments passed to other methods.}
}
\value{
A matrix of predicted risks with subjects in rows and time points in columns.
}
\description{
This function predicts the absolute risk for a specified cause from a
\code{CompRisk} model object, compatible with the \code{riskRegression} package.
}
