% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predictRisk.iCoxBoost}
\alias{predictRisk.iCoxBoost}
\title{Predict Cumulative Incidence for an iCoxBoost Object}
\usage{
\method{predictRisk}{iCoxBoost}(object, newdata, times, cause, ...)
}
\arguments{
\item{object}{A model object of class \code{iCoxBoost}.}

\item{newdata}{A \code{data.frame} containing the predictor variables.}

\item{times}{A numeric vector of time points at which to predict risk.}

\item{cause}{The specific event type for which to predict the CIF.}

\item{...}{Additional arguments passed to other methods.}
}
\value{
A matrix of predicted cumulative incidences with subjects in rows
and time points in columns.
}
\description{
This function predicts the cumulative incidence function (CIF) for a specified
cause from an \code{iCoxBoost} model object.
}
