% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predictRisk.twoCSlassos}
\alias{predictRisk.twoCSlassos}
\title{Predict Absolute Risk for a twoCSlassos Object}
\usage{
\method{predictRisk}{twoCSlassos}(
  object,
  newdata,
  times,
  cause,
  lambdavecs = NULL,
  indices = NULL,
  ...
)
}
\arguments{
\item{object}{A fitted object of class \code{twoCSlassos}.}

\item{newdata}{A data frame containing the covariate values for which
to predict.}

\item{times}{A vector of time points at which to predict the
absolute risk.}

\item{cause}{The cause of interest for which to predict risk.}

\item{lambdavecs}{A list of lambda vectors. If \code{NULL}, defaults to
\code{object$lambdas}.}

\item{indices}{An integer vector specifying which index from each
lambda vector to use for prediction. If \code{NULL}, defaults to the
last index (the smallest lambda) of each path.}

\item{...}{Additional arguments passed to
\code{predictEventProb.twoCSlassos}.}
}
\value{
A matrix of predicted absolute risks, with rows corresponding
to \code{newdata} and columns to \code{times}.
}
\description{
This function is the S3 method for \code{\link[riskRegression]{predictRisk}}
for an object of class \code{twoCSlassos}. It serves as a wrapper
around \code{predictEventProb.twoCSlassos}.
}
\details{
This function allows the model to be used with the
\code{\link[riskRegression]{Score}} function for evaluating Brier scores
and other metrics. It selects the last lambda value from each path
Details:
This function allows the model to be used with the
\code{\link[riskRegression]{Score}} function for evaluating Brier scores
and other metrics. It selects the last lambda value from each path
by default.
}
