% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predictSurvProb.oneCSlasso}
\alias{predictSurvProb.oneCSlasso}
\title{Predict Survival Probabilities for a oneCSlasso Object}
\usage{
\method{predictSurvProb}{oneCSlasso}(object, newdata, times, lambdavec, index, ...)
}
\arguments{
\item{object}{A fitted object of class \code{oneCSlasso}.}

\item{newdata}{A data.frame for which to predict survival.}

\item{times}{A numeric vector of times to predict at.}

\item{lambdavec}{The lambda vector used for the fit.}

\item{index}{The specific index of the \code{lambdavec} to use
for prediction.}

\item{...}{Not used.}
}
\value{
A matrix of survival probabilities (rows=newdata, cols=times).
}
\description{
S3 method for \code{predictSurvProb} for an object of
class \code{oneCSlasso}.
}
