% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.r
\name{prepare_penalty_params}
\alias{prepare_penalty_params}
\title{Prepare Penalty Parameters for the Fitting Function}
\usage{
prepare_penalty_params(regularization, lambda, alpha)
}
\arguments{
\item{regularization}{A string, either 'elastic-net' or 'SCAD'.}

\item{lambda}{The primary shrinkage parameter.}

\item{alpha}{The mixing parameter for elastic-net or shape parameter for SCAD.}
}
\value{
A list containing the calculated lambda1 and lambda2 values.
}
\description{
A helper function to calculate the lambda parameters required by the mtool
fitting function based on the chosen regularization method.
}
