% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{print.cbSCRIP.path}
\alias{print.cbSCRIP.path}
\title{Print a cbSCRIP.path object}
\usage{
\method{print}{cbSCRIP.path}(x, ...)
}
\arguments{
\item{x}{An object of class \code{cbSCRIP.path}, the result of \code{path_cbSCRIP}.}

\item{...}{Additional arguments passed to other methods.}
}
\value{
The original object \code{x}, invisibly.
}
\description{
Print a cbSCRIP.path object
}
