% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.r
\name{print.cbSCRIP}
\alias{print.cbSCRIP}
\title{Print method for cbSCRIP objects}
\usage{
\method{print}{cbSCRIP}(x, ...)
}
\arguments{
\item{x}{An object of class 'cbSCRIP'.}

\item{...}{Additional arguments.}
}
\description{
Print method for cbSCRIP objects
}
