% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{print.MNlogisticFit}
\alias{print.MNlogisticFit}
\title{Print method for MNlogisticFit objects}
\usage{
\method{print}{MNlogisticFit}(x, ...)
}
\arguments{
\item{x}{An object of class 'MNlogisticFit'.}

\item{...}{Additional arguments.}
}
\description{
Print method for MNlogisticFit objects
}
