% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{refit_cbSCRIP}
\alias{refit_cbSCRIP}
\title{Refit Models Along a Regularization Path}
\usage{
refit_cbSCRIP(object, ...)
}
\arguments{
\item{object}{A fitted object of class}

\item{...}{Additional arguments (currently unused).}

\item{coeffs}{A character string specifying the type of coefficients to return.}
}
\value{
An object of class \code{cbSCRIP.path}, where the \code{coefficients} list
contains the refitted coefficients. A new element, \code{refitted_models}, is
also added, containing the full model objects from each refit.
}
\description{
For each lambda in a \code{cbSCRIP.path} object, this function identifies the
selected variables and refits a standard (unpenalized) \code{casebase} model
using only that subset.
}
