% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoLasso.R
\name{two.i.CSlassos}
\alias{two.i.CSlassos}
\title{Fit Two Cause-Specific Lasso Models with Internal Lambda Grid}
\usage{
two.i.CSlassos(data, nlambda = 100, lambda_min_ratio = 1e-04, ...)
}
\arguments{
\item{data}{A data.frame containing survival time, status, and covariates.}

\item{nlambda}{The number of lambda values to use.}

\item{lambda_min_ratio}{The ratio of smallest to largest lambda.}

\item{...}{Additional arguments passed to \code{twoCSlassos}.}
}
\value{
An object of class \code{twoCSlassos}.
}
\description{
A helper function that creates a default lambda grid and then
calls \code{twoCSlassos}.
}
