% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoLasso.R
\name{two.selpen.CSlassos}
\alias{two.selpen.CSlassos}
\title{Select Penalties for twoCSlassos}
\usage{
two.selpen.CSlassos(object, ind)
}
\arguments{
\item{object}{A fitted \code{twoCSlassos} object.}

\item{ind}{A vector of indices for the selected lambdas.}
}
\value{
An object of class \code{two.selpen.CSlassos}.
}
\description{
A helper function to store a \code{twoCSlassos} object along with
a selected set of indices.
}
