% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.r
\name{unstandardize_coefficients}
\alias{unstandardize_coefficients}
\title{Unstandardize Coefficients}
\usage{
unstandardize_coefficients(coefficients, scaler, col_names)
}
\arguments{
\item{coefficients}{A matrix or vector of standardized coefficients.}

\item{scaler}{A list with \code{center} and \code{scale} components.}

\item{col_names}{Original column names for the covariates.}
}
\value{
The unstandardized coefficients.
}
\description{
Transforms coefficients from the standardized scale back to the original scale.
}
