% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{create_lambda_grid}
\alias{create_lambda_grid}
\title{Generate a Lambda Grid for Regularization}
\usage{
create_lambda_grid(
  cb_data,
  lambda,
  nlambda,
  lambda_max,
  lambda.min.ratio,
  regularization,
  alpha,
  n_unpenalized,
  ...
)
}
\value{
A numeric vector of lambda values.
}
\description{
Generate a Lambda Grid for Regularization
}
