% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv_cbSCRIP}
\alias{cv_cbSCRIP}
\title{Cross-Validation for Penalized Multinomial Case-Base Models}
\usage{
cv_cbSCRIP(
  formula,
  data,
  regularization = "elastic-net",
  cb_data = NULL,
  alpha = NULL,
  lambda = NULL,
  nfold = 5,
  nlambda = 50,
  ncores = parallel::detectCores()/2,
  n_unpenalized = 2,
  ratio = 25,
  ratio_event = 1,
  lambda_max = NULL,
  lambda.min.ratio = NULL,
  warm_start = T,
  ...
)
}
\value{
An object of class \code{cb.cv} containing the results.
}
\description{
Cross-Validation for Penalized Multinomial Case-Base Models
}
