% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{find_lambda_max}
\alias{find_lambda_max}
\title{Find the Maximum Lambda (lambda_max)}
\usage{
find_lambda_max(cb_data, n_unpenalized, n_event_types, ncores, fit_fun, ...)
}
\arguments{
\item{...}{Other arguments passed to find_lambda_max.}
}
\value{
The estimated lambda_max value.
}
\description{
Find the Maximum Lambda (lambda_max)
}
