% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{fit_cb_model}
\alias{fit_cb_model}
\title{Fit a Penalized Multinomial Model on Case-Base Data}
\usage{
fit_cb_model(
  cb_data,
  regularization = c("elastic-net", "SCAD"),
  lambda,
  alpha = NULL,
  fit_fun = mtool::mtool.MNlogistic,
  param_start = NULL,
  standardize = TRUE,
  all_event_levels = NULL
)
}
\arguments{
\item{cb_data}{A case-base dataset from \code{create_cb_data}.}

\item{regularization}{A string, either 'elastic-net' or 'SCAD'.}

\item{lambda}{The primary shrinkage parameter.}

\item{alpha}{The mixing/shape parameter. See \code{prepare_penalty_params}.}

\item{fit_fun}{The fitting function from \code{mtool} to use.}

\item{param_start}{Optional starting values for the coefficients.}

\item{standardize}{Logical. If TRUE, covariates are scaled to have mean 0 and SD 1.}

\item{unpen_cov}{Integer. The number of leading covariates to leave unpenalized.}
}
\value{
The fitted model object from the specified \code{fit_fun}.
}
\description{
A wrapper function to fit a penalized model (Elastic-Net or SCAD) using
an optimizer from the \code{mtool} package.
}
