% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{MNlogisticFISTA}
\alias{MNlogisticFISTA}
\title{Multinomial Logistic Regression with FISTA}
\usage{
MNlogisticFISTA(
  X,
  Y,
  offset,
  N_covariates,
  regularization = "l1",
  transpose = FALSE,
  lambda1,
  lambda2 = 0,
  lambda3 = 0,
  pos = FALSE,
  tolerance = 1e-05,
  maxit = 1000,
  ncores = -1,
  lr_adj = 1,
  learning_rate = 1,
  group_id = NULL,
  group_weights = NULL,
  groups = NULL,
  groups_var = NULL,
  own_variables = NULL,
  N_own_variables = NULL,
  param_start = NULL,
  verbose = FALSE,
  save_history = FALSE,
  lam1_prev = 0
)
}
\arguments{
\item{X}{Input matrix.}

\item{Y}{Response vector.}

\item{offset}{Offset vector.}

\item{N_covariates}{Number of covariates.}

\item{regularization}{Regularization type.}

\item{transpose}{Transpose X.}

\item{lambda1}{Primary penalty parameter.}

\item{lambda2}{Secondary penalty parameter.}

\item{lambda3}{Tertiary penalty parameter.}

\item{pos}{Positivity constraint.}

\item{tolerance}{Convergence tolerance.}

\item{maxit}{Maximum iterations.}

\item{ncores}{Number of cores.}

\item{lr_adj}{Learning rate adjustment.}

\item{learning_rate}{Learning rate.}

\item{group_id}{Group IDs.}

\item{group_weights}{Group weights.}

\item{groups}{Groups matrix.}

\item{groups_var}{Groups variable matrix.}

\item{own_variables}{Own variables.}

\item{N_own_variables}{Number of own variables.}

\item{param_start}{Starting parameters.}

\item{verbose}{Verbosity.}

\item{save_history}{Save history.}

\item{lam1_prev}{Previous \code{lambda1} on the path, used by the sequential
strong rules to pre-screen the active set. Ignored on a cold start.}
}
\value{
A list with model results.
}
\description{
Fits a multinomial logistic regression model with L1/L2 or SCAD regularization
using FISTA (Fast Iterative Shrinkage-Thresholding Algorithm).
}
