% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.r
\name{print.cbSCRIP}
\alias{print.cbSCRIP}
\title{Print a cbSCRIP object}
\usage{
\method{print}{cbSCRIP}(x, ...)
}
\arguments{
\item{x}{An object of class \code{cbSCRIP}, the result of \code{fit_cb_model}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The original object \code{x}, invisibly.
}
\description{
Print a cbSCRIP object
}
