% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{run_cv_fold}
\alias{run_cv_fold}
\title{Run a Single Fold of Cross-Validation}
\usage{
run_cv_fold(
  fold_indices,
  cb_data,
  lambdagrid,
  all_event_levels,
  regularization,
  alpha,
  n_unpenalized = 2,
  warm_start = TRUE,
  update_f = NULL,
  optimizer = "CCD",
  ...
)
}
\arguments{
\item{fold_indices}{Indices for the validation fold.}

\item{cb_data}{Case-base data.}

\item{lambdagrid}{Vector of lambda values.}

\item{all_event_levels}{Vector of event levels.}

\item{regularization}{Type of regularization.}

\item{alpha}{The alpha value.}

\item{n_unpenalized}{Number of unpenalized predictors.}

\item{warm_start}{Logical, whether to use warm starts.}

\item{update_f}{A progress update function.}

\item{...}{Additional arguments for \code{fit_cb_model}.}
}
\value{
A list of multinomial deviances and non-zero counts for the fold.
}
\description{
Run a Single Fold of Cross-Validation
}
