% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{run_cv_fold}
\alias{run_cv_fold}
\title{Run a Single Fold of Cross-Validation}
\usage{
run_cv_fold(
  fold_indices,
  cb_data,
  lambdagrid,
  all_event_levels,
  regularization,
  alpha,
  n_unpenalized = 2,
  warm_start = T,
  update_f = NULL,
  ...
)
}
\value{
A vector of multinomial deviances for the fold.
}
\description{
Run a Single Fold of Cross-Validation
}
