% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.r
\name{standardize_cb_data}
\alias{standardize_cb_data}
\title{Standardize Case-Base Data}
\usage{
standardize_cb_data(cb_data)
}
\arguments{
\item{cb_data}{A case-base data list.}
}
\value{
A list containing the standardized \code{cb_data} and the \code{scaler} object.
}
\description{
Centers and scales the covariates in a case-base dataset.
}
